server/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
waitress>=2.1.2
